Factor out the filepicker
